app-health-back/src/@api/app-health/authorization-interface/resolvers/app-health-get-authorization-interfaces.resolver.ts
import { AppHealthGetAuthorizationInterfacesHandler } from '@api/app-health/authorization-interface';
import { AppHealthAuthorizationInterface } from '@api/graphql';
import { QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Query, Resolver } from '@nestjs/graphql';

@Resolver()
export class AppHealthGetAuthorizationInterfacesResolver
{
    constructor(
        private readonly handler: AppHealthGetAuthorizationInterfacesHandler,
    ) {}

    @Query('appHealthGetAuthorizationInterfaces')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
    ): Promise<AppHealthAuthorizationInterface[]>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
        );
    }
}
